THOR-1331 Updates based on CR.
